requirements.txt
boto==2.49.0
boto3==1.9.124
botocore==1.12.124
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-formtools==2.1
django-storages==1.7.1
django-tinymce==2.8.0
django-widget-tweaks==1.4.3
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==2.8.0
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
